Boss fixed

+++ src/TheFindingOfIZack/Entities/Boss.java
@@ -17,8 +17,8 @@ public class Boss extends Enemy {
      */
     public static int size = 120;
 
-    private int armour = 500;
-    private int MAX_ARMOUR = 500;
+    private double armour = 500;
+    private double MAX_ARMOUR = 500;
 
     /**
      * Constructor for the boss
